msm: ipa: ipa_usb: remove SUSPEND_IN_PROGRESS state
Remove the usage of SUSPEND_IN_PROGRESS state in IPA USB.
This state was introduced to allow USB host suspend without setting
USB controller to LPM. As this scenario is not a requirement,
SUSPEND_IN_PROGRESS is not supported anymore. Instead IPA USB will go to
full suspend state.

Change-Id: I7224144fcf6cb70a030657aa59cb858fa3d410d8
CRs-Fixed: 2054613
Acked-by: Ady Abraham <adya@qti.qualcomm.com>
Signed-off-by: Skylar Chang <chiaweic@codeaurora.org>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>