usb: dwc3: Enable evicting endpoint cache after flow control
In 3.00a and 3.20a synopsys release a performance enhancemnet
was done to keep the Non Stream capable bulk IN endpoint in cache
after flow control. This enhancement causes host mode concurrency
issue in which xhci command timeout is happening when mutliple mass
storage devices is connected to the device using HUB and
simultaneous file transfers are happening to mass storage devices.
This failure is caused because if one endpoint flow controlled,
other endpoints will not be serviced right. This change sets
EnableEpCacheEvict bit which will disable the enhancement.

Change-Id: Iafb72f75b81b1bcbe8c5e38ebd6fe40905670e7a
Signed-off-by: Pratham Pratap <prathampratap@codeaurora.org>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>