usb: gadget: Add support for IPC driver
IPC messages are to be exchanged between device and host using
the IPC router platform driver on the device side. For this
driver to use USB as transport layer, add support for function
driver which will configure Bulk IN and OUT endpoints and expose
proper callbacks to router driver for read and write operations.

Change-Id: I350624ccd24a6c6fd4bda5ea91379fb426fae742
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>