power: qpnp-fg-gen3: adjust ki coefficient for full soc dynamically
In certain conditions, after the battery SOC reaches 100%, SOC is
not getting updated during discharging. Based on the hardware
recommendation, adjust the Ki coefficient for full SOC so that
SOC can be updated faster during discharging.

While at it, remove prev_charge_status variable which is not
used really.

CRs-Fixed: 2089555
Change-Id: I48b3494165396d36ce5b4212a0464d4ea87ee9a4
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: Sahil Chandna <chandna@codeaurora.org>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>