usb: gadget: f_ipc: Prevent the unregister call from getting stuck
Set the state to IPC_DISCONNECTED and wake up the
write/read wait queues before calling
platform_device_unregister to prevent the call from getting
stuck. The call gets stuck because the IPC export is unable
to unregister from the core if it is stuck waiting in
the ipc_write call.

Change-Id: I4fc04548c6927c1adc0cd746dbd07a02712d0d1e
Signed-off-by: Gustavo Solaira <gustavos@codeaurora.org>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>