drivers: iio: imu: Fix NULL pointer dereference in bmi160 driver
The IIO buffer is dereference to NULL when user space application
trying to change the parameters of iio buffer while driver pushing data
to iio buffer. The issue is solved by using proper initialization of
iio buffer and using standard api to push the data to iio buffer so that
driver will do proper conditon checking before changing the
parameters of buffer and before pushing the data.

Change-Id: I6c9b015fc2915c0d6f635e488305290a36eb9221
Signed-off-by: Puneet Yatnal <puneet@codeaurora.org>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>