kernel: cgroup: boost to max whenever a zygote-forked process becomes a top-app
Boost to the max for 128ms whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

original idea from sultanxda

Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>
Signed-off-by: ronaxdevil <pratabidya.007@gmail.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>